+++ packages/forge/src/lib/core/index.ts
@@ -1,5 +1,6 @@
 export * from './base/index.js';
 export * from './configuration/index.js';
+export * from './decorators/index.js';
 export * from './delegates/index.js';
 export * from './utils/index.js';
 export * from './mask/index.js';